Fix bug with menu at 600px exactly
At 600px, the menu was still recieving the mobile stylings, but was
located in the desktop location. This is due to the menu having been
created and styled at desktop first instead of mobile first.

This should be fixed.

This current fix adjusts the max-width down such that it equates to 1px.